+++ Weavers/MvxMainThread.Fody/ModuleWeaver.ShouldWeave.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Linq;
+﻿using System.Linq;
 using Mono.Cecil;
 
 public partial class ModuleWeaver
